refactor: add custom hook for fileTree & fileTabs & dependency
